1.1.0.4 Update
Removed 'Help' and 'Lock/Unlock' buttons from main dashboard
Created QAction button for 'Help' on the toolbar
Created QAction button for 'Lock/Unlock' on the toolbar
Updated some art assets

Fixed Minimize menu by chaning it to a button and using the minimize
preferences in the option screen

BUG: Only 'Minimize To Tray' works on Debian based distros: QT4 bug with
Qt::WindowMinimized.  This should be resolved if building with QT5.